Change all activities to respond to theme changes in onCreate and onResume. This also refactors a few of the theme helper functions into the ThemeUtils class from the AndroidUtils class. Finally, this also removes the explicit UI colors from the themes that were using the reference notation (?). Those colors weren't working and they will be replaced soon with more intentional colors anyway.
