+++ app/src/main/java/org/moire/opensudoku/gui/FileListActivity.java
@@ -37,6 +37,7 @@ import android.widget.SimpleAdapter.ViewBinder;
 import android.widget.TextView;
 
 import org.moire.opensudoku.R;
+import org.moire.opensudoku.utils.ThemeUtils;
 
 import java.io.File;
 import java.text.DateFormat;
@@ -67,8 +68,12 @@ public class FileListActivity extends ListActivity {
     private List<Map<String, Object>> mList;
     private Context mContext = this;
 
+    private int mThemeId = 0;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
+        ThemeUtils.setThemeFromPreferences(this);
+        mThemeId = ThemeUtils.getThemeResourceIdFromPreferences(this);
         super.onCreate(savedInstanceState);
 
         setContentView(R.layout.file_list);
@@ -133,6 +138,16 @@ public class FileListActivity extends ListActivity {
     @Override
     protected void onDestroy() {
         super.onDestroy();
+    }
+
+    @Override
+    protected void onResume() {
+        super.onResume();
+
+        int newThemeId = ThemeUtils.getThemeResourceIdFromPreferences(this);
+        if (newThemeId != mThemeId) {
+            recreate();
+        }
     }
 
     @Override
