+++ app/src/main/java/org/moire/opensudoku/utils/AndroidUtils.java
@@ -35,37 +35,6 @@ public class AndroidUtils {
         return list.size() > 0;
     }
 
-    public static int getThemeResourceIdFromString(String theme){
-        switch (theme) {
-            case "default":
-                return R.style.Theme_Default;
-            case "opensudoku":
-                return R.style.Theme_OpenSudoku;
-            case "paper":
-                return R.style.Theme_Paper;
-            case "graphpaper":
-                return R.style.Theme_GraphPaper;
-            case "light":
-                return R.style.Theme_Light;
-            case "paperlight":
-                return R.style.Theme_PaperLight;
-            case "graphpaperlight":
-                return R.style.Theme_GraphPaperLight;
-            case "highcontrast":
-                return R.style.Theme_HighContrast;
-            case "invertedhighcontrast":
-                return R.style.Theme_InvertedHighContrast;
-            default:
-                return R.style.Theme_Default;
-        }
-    }
-
-    public static void setThemeFromPreferences(Context context) {
-        SharedPreferences gameSettings = PreferenceManager.getDefaultSharedPreferences(context);
-        String theme = gameSettings.getString("theme", "default");
-        context.setTheme(getThemeResourceIdFromString(theme));
-    }
-
     /**
      * Returns version code of OpenSudoku.
      *
